rename chain.rs -> stream_scan.rs
