+++ src/stream/src/from_proto/mod.rs
@@ -18,7 +18,6 @@ mod agg_common;
 mod append_only_dedup;
 mod barrier_recv;
 mod batch_query;
-mod chain;
 mod dml;
 mod dynamic_filter;
 mod eowc_over_window;
@@ -43,6 +42,7 @@ mod sink;
 mod sort;
 mod source;
 mod stateless_simple_agg;
+mod stream_scan;
 mod temporal_join;
 mod top_n;
 mod union;
@@ -58,7 +58,6 @@ use risingwave_storage::StateStore;
 use self::append_only_dedup::*;
 use self::barrier_recv::*;
 use self::batch_query::*;
-use self::chain::*;
 use self::dml::*;
 use self::dynamic_filter::*;
 use self::eowc_over_window::*;
@@ -83,6 +82,7 @@ use self::sink::*;
 use self::sort::*;
 use self::source::*;
 use self::stateless_simple_agg::*;
+use self::stream_scan::*;
 use self::temporal_join::*;
 use self::top_n::*;
 use self::union::*;
