+++ src/stream/src/from_proto/stream_scan.rs
